updated references to daemon after moving it to extras
