+++ lbrynet/extras/daemon/auth/client.py
@@ -4,7 +4,7 @@ import logging
 from urllib.parse import urlparse
 
 from lbrynet import conf
-from lbrynet.daemon.auth.keyring import Keyring, APIKey
+from .keyring import Keyring, APIKey
 
 log = logging.getLogger(__name__)
 USER_AGENT = "AuthServiceProxy/0.1"
