+++ tests/unit/lbrynet_daemon/test_Downloader.py
@@ -9,8 +9,8 @@ from lbrynet.core.Error import DownloadDataTimeout, DownloadSDTimeout
 from lbrynet.core.StreamDescriptor import StreamDescriptorIdentifier
 from lbrynet.core.BlobManager import DiskBlobManager
 from lbrynet.core.RateLimiter import DummyRateLimiter
-from lbrynet.daemon import Downloader
-from lbrynet.daemon.ExchangeRateManager import ExchangeRateManager
+from lbrynet.extras.daemon import Downloader
+from lbrynet.extras.daemon import ExchangeRateManager
 from lbrynet.database.storage import SQLiteStorage
 from lbrynet.dht.peerfinder import DummyPeerFinder
 from lbrynet.file_manager.EncryptedFileStatusReport import EncryptedFileStatusReport
