+++ tests/unit/lbrynet_daemon/test_claims_comparator.py
@@ -1,6 +1,6 @@
 import unittest
 
-from lbrynet.daemon.Daemon import sort_claim_results
+from lbrynet.extras.daemon.Daemon import sort_claim_results
 
 
 class ClaimsComparatorTest(unittest.TestCase):
